+++ mrb_doc/core/window.rb
@@ -7,15 +7,6 @@ def set_window_monitor(monitor)
   nil
 end
 
-# Gets the refresh rate of the specified monitor.
-# @param monitor [Integer]
-# @return [Integer]
-def get_monitor_refresh_rate(monitor)
-  # mrb_get_monitor_refresh_rate
-  # src/mruby_integration/core/window.cpp
-  60
-end
-
 # Gets the scale of the window.
 # @return [Vector2]
 def get_window_scale_dpi