+++ src/mruby_integration/models/monitor.cpp
@@ -122,6 +122,20 @@ mrb_Monitor_height(mrb_state* mrb, mrb_value self) -> mrb_value
   return mrb_int_value(mrb, GetMonitorHeight(monitor->id));
 }
 
+auto
+mrb_Monitor_refresh_rate(mrb_state* mrb, mrb_value self) -> mrb_value
+{
+  EXIT_UNLESS_WINDOW_READY(
+    "You must call Window.open before Monitor#refresh_rate");
+
+  Monitor* monitor;
+
+  Data_Get_Struct(mrb, self, &Monitor_type, monitor);
+  mrb_assert(monitor != nullptr);
+
+  return mrb_int_value(mrb, GetMonitorRefreshRate(monitor->id));
+}
+
 void
 append_models_Monitor(mrb_state* mrb)
 {
@@ -141,6 +155,11 @@ append_models_Monitor(mrb_state* mrb)
     mrb, Monitor_class, "width", mrb_Monitor_width, MRB_ARGS_NONE());
   mrb_define_method(
     mrb, Monitor_class, "height", mrb_Monitor_height, MRB_ARGS_NONE());
+  mrb_define_method(mrb,
+                    Monitor_class,
+                    "refresh_rate",
+                    mrb_Monitor_refresh_rate,
+                    MRB_ARGS_NONE());
 
   load_ruby_models_monitor(mrb);
 }